Move test/train_mapper_class fns to astrodet; update hsc branch within them
